pyproject.toml
[tool.poetry]
authors = ["Steven Biro <steven.biro@mail.utoronto.ca>"]
description = "A bot for the FFR discord server"
license = "GPLv3"
name = "ffr-bot"
version = "2.0.0-alpha"

[tool.poetry.dependencies]
asyncio = "^3.4.3"
discord-ext-typed-commands = "^1.0.3"
"discord.py" = "^1.6.0"
"discord.py-stubs" = "^1.6.0"
pymongo = "^3.11.3"
pymongo-stubs = "^0.1.0"
python = "^3.8"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.812"
poethepoet = "^0.10.0"
watchdog = "^2.0.2"

[tool.poe.tasks]
dev = {cmd = "scripts/dev.sh", help = "lauches mongod with the data in .data/<arg> or .data/db1 if nothing is passed in, then runs ffrbot.dev, which watches for file changes and runs ffrbot"}
mypy = {cmd = "poetry run mypy ffrbot", help = "runs mypy on ffrbot"}
mypy-watch = {cmd = "scripts/mypy-watch.sh", help = "watches the supplied directory or file with mypy, defaults to ffrbot when nothing is passed in"}
start = {cmd = "poetry run python3 -m ffrbot", help = "runs ffrbot"}

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
